Move Draw::draw_frame to new trait DrawShaded within kas_wgpu
