+++ kas-wgpu/src/draw/round_pipe.rs
@@ -8,7 +8,7 @@
 use std::f32::consts::FRAC_PI_2;
 use std::mem::size_of;
 
-use kas::draw::*;
+use crate::draw::{Colour, Quad, Vec2};
 use kas::geom::Size;
 
 use super::Rgb;
